Accounts list improvements, spinner service & dialogs module (#10)
* Include material layout

* Basic list of accounts.

* Add shared modules for common dependencies.

* Install Angular2 Material Flex Layout.

* Print account information in account list.

* Account actions.

* Build redirect url after login dinamically.

* Place the remove account function in the correct component.

* Remove real data.

* List of accounts as observable subject.

* New module to display dialogs and implementation for confirm dialogs.

* Include Material Icons.

* Improve account list display.

* Show confirm dialog when disconnecting an account.

* Added snackbar  to account actions.

* Add spinner to account actions.

* Add spinner to account actions.

* Change spinner color.

+++ src/app/accounts/account/account.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Account } from '../account';
 import { AccountService } from '../account.service';
+import { MdSnackBar } from '@angular/material';
+import { DialogsService } from '../../shared/dialogs/dialogs.service';
+import { SpinnerService } from '../../shared/providers/spinner.service';
+import 'rxjs/add/operator/do';
 
 @Component({
   selector: 'app-account',
@@ -11,7 +15,10 @@ export class AccountComponent implements OnInit {
 
   @Input() account: Account;
 
-  constructor(private accountService: AccountService) { }
+  constructor(private accountService: AccountService,
+              private dialogService: DialogsService,
+              private spinnerService: SpinnerService,
+              private snackBar: MdSnackBar) { }
 
   ngOnInit() {
   }
@@ -21,8 +28,15 @@ export class AccountComponent implements OnInit {
    * @param account The account to disconnect.
    */
   removeAccount(account: Account): void {
-    this.accountService.removeAccount(account)
-        .subscribe(() => console.log('account removed'));
+
+    this.dialogService.confirm('Confirm Remove', 'Are you sure you want to disconnect this account?')
+        .filter(res => res)
+        .do(() => this.spinnerService.showSpinner())
+        .subscribe(() => {
+          return this.accountService.removeAccount(account)
+                     .do(() => this.spinnerService.hideSpinner())
+                     .subscribe(() => this.snackBar.open('Account successfully removed', null, { duration: 2000 }))
+        });
   }
 
 }
